Adopt commit 25b942f from master: Merge pull request #340 from abeels/warnings-gcc-5.3
commit 25b942f3240e10b3b0e2a3e57c6ae3d75d512878
Merge: 5f4cc73 83f4567
Author: Alexander Beels <abeels@bloomberg.net>
Date:   Wed Apr 20 19:20:07 2016 -0400

    Merge pull request #340 from abeels/warnings-gcc-5.3

    Silence conversion warnings

commit 83f456759fc7bae348f96263dcb0dd0492f7d255
Author: Alexander Beels <abeels@bloomberg.net>
Date:   Wed Apr 20 18:50:28 2016 -0400

    Fix remaining use of 'ssize_t'

commit fbed9a6db7429681d4c6cb7a241aed72243ab923
Author: Alexander Beels <abeels@bloomberg.net>
Date:   Wed Apr 20 18:38:41 2016 -0400

    Treat ssize_t as Int64.

commit dd767a0e62f038004534e601f6791174f69eafbd
Author: Alexander Beels <abeels@bloomberg.net>
Date:   Tue Apr 19 16:32:12 2016 -0400

    Fix use of 'addCharacters' in test driver.

commit a8b209714863949b9b40e8d16c7ddf838752cb12
Author: Alexander Beels <abeels@bloomberg.net>
Date:   Tue Apr 19 15:23:08 2016 -0400

    Silence conversion warnings (gcc-5.3 64-bit).

+++ groups/bal/balb/balb_testmessages.h
@@ -5086,7 +5086,9 @@ const bsl::string& CustomString::toString() const
 inline
 int Enumerated::fromString(Value *result, const bsl::string& string)
 {
-    return fromString(result, string.c_str(), string.length());
+    return fromString(result,
+                      string.c_str(),
+                      static_cast<int>(string.length()));
 }
 
 inline
